Cleaning up Composite Structure Diagram Test/Interface Layer
adding the interface is done twice, once via adding existing elements
and a second time via drag and drop from the model explore (depending on
the test execution, it can happen that this dropping it onto a component
and adding extra connections).
Cleaning up, and removing the drag and drop action.